cloud_function/requirements.txt
scikit-learn
joblib
google-cloud-storage
